[YUNIKORN-231] Update container state message (#21)
Define a message to communicate scheduling state for a container back to
a shim. This provides a shim with the correct state of a container.

The first usage is on K8s with the auto scaler. This removes the need
for the shim to assume a pod state to help the auto scaler grow the
number of nodes.

Fixes: #21