[YUNIKORN-374] Update gRPC version to 1.26.0 (#25)
Update the gRPC to 1.26.0 in and bring it inline with the version of
gRPC of the core repo. This includes a version change of protobuf.
The protobuf version specified in the go.mod file is passed through into
the Makefile for the generation removing the hard coded version from the
Makefile.

Fixes: #25