[YUNIKORN-114] Shallow clone of protobuf repo (#28)
Make a shallow clone of the protobuf repository of just the tag
specified in the go.mod file.
To allow changes of the version without intervention a re-clone is
triggered if needed. This does rely on a cleanup which fails if the root
of the repo contains hidden file with a single character as the name.

Fixes: #28