[YUNIKORN-760] Fix build failures on clean checkout
The update of the proto compiler caused the build to fail on a clean
checkout.
The new protoc uses a different plugin setup and stricter rules around
the proto content. gRPC generations has been removed at this point as it
is not used. Simplified the protoc command line.
Make clean and clobber consistentcy changes.