[YUNIKORN-815] Fix scheduler interface makefile issues
Downgrade gRPC version, because the shim repo depends on K8s 1.16, the etcd is not
compatible with gRPC 1.32, see: https://github.com/etcd-io/etcd/issues/12124. As a
result, we are stick to gRPC 1.26.0 for now, unless the etcd updated to a version
that is compatible with the newer gRPC versions. Accordingly, we need to use the plugin
"--go_out=plugins=grpc" option to generate the gRPC code, otherwise it will generate
some annotations that is not compatible with gRPC 1.26.0.