[YUNIKORN-907] Message and interface refactor (#56)
The current interface wraps a number of smaller dedicated messages into
one large update message. This provides a simple interface to implement
with just one message to be handled. The complexity is hidden inside the
unwrapping of the message.
This change reverses the process by sending simple dedicated messages
and removing teh wrapping. Nodes, applications, allocations and
configuration are processed in their own functions.

The plugin is now defined as part of the SI and not in the core to start
breaking the dependency of the shim on the core.

Cleanup of spurious space, tabs and code re-generated based on the
latest changes.

Co-authored-by: Wilfred Spiegelenburg <wilfreds@apache.org>

Fixes: #56